Refactor rules fetcher into modules
