+++ src/util/rules_fetcher/__main__.py
@@ -1,0 +1,9 @@
+import logging
+
+from . import _run_as_script
+
+if __name__ == "__main__":
+    logging.basicConfig(
+        level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+    )
+    _run_as_script()
